Handle switches to fullscreen.
The Y position of all the views needs to be updated every time
the system goes in and out of fullscreen mode so that the controls
stay aligned to the bottom of the screen.

SystemUiVisiblityChangeListeners aren't being triggered reliably
so instead build an equivalent using a frame layout that seems to
be triggered every time.

Test: Built a fullscreen activity and tested moving in and out of
      full screen.
Change-Id: Ic976e4c6b2abdb8afe8dced2684fccbbc9dea53e