Make Hvac UI common across all users.
The hvac ui and its backing service do not store any per user
data so we can save on ram by making it so only one copy of
the service is ever instantiated. We do this by marking the
services as single instance. This also prevents a bug where
we see 2 copies of the hvac ui when switching users.

Bug: 36454400
Test: Booted, switched users, verified that there weren't 2
      copies of the HVAC ui.

Change-Id: I220a8f906a8271a57fce0b7923f7826e81d6f1de